% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeedata-package.r
\docType{package}
\name{coffeedata}
\alias{coffeedata}
\alias{coffeedata-package}
\title{coffeedata.}
\description{
Every time coffee was consumed by Roel, a time stamp was added.
The data contains individual observations, every row is a seperate observation.
}

